requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
Flask>=2.3.2
aiohttp>=3.8.4
waitress>=2.1.2
asyncio